duhacks/multi_agent_code_reviewer/requirements.txt
boto3==1.34.0
botocore==1.34.0
google-generativeai==0.3.0
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
GitPython==3.1.40
python-dotenv==1.0.0
